+++ src/tools/viewer/widgets/LVRReconstructionEstimateNormalsDialog.cpp
@@ -110,7 +110,7 @@ void LVREstimateNormalsDialog::estimateNormals()
         for(int i = 0; i < numPoints; i++)
         {
             // Create search tree
-            vector< ulong > indices;
+            vector< int > indices;
             vector< float > distances;
 
             Vertex<float> vertex(points[3 * i], points[3 * i + 1], points[3 * i + 2]);
